auth ✔

+++ src/enrollment/enrollment.controller.ts
@@ -6,20 +6,25 @@ import {
   Patch,
   Param,
   Delete,
+  UseGuards,
+  UnauthorizedException,
 } from '@nestjs/common';
 import { EnrollmentService } from './enrollment.service';
 import { CreateEnrollmentDto } from './dto/create-enrollment.dto';
+import { AuthGuard } from 'src/auth/guards/auth.guard';
+import { CurrentUser } from 'src/common/decorators/current-user.decorator';
 
 @Controller('enrollments')
 export class EnrollmentController {
   constructor(private readonly enrollmentService: EnrollmentService) {}
 
+  @UseGuards(AuthGuard)
   @Post()
   async create(@Body() data: CreateEnrollmentDto) {
     const result = await this.enrollmentService.create(data);
     return { message: 'enrollment issued successfuly !', data: result };
   }
-
+  @UseGuards(AuthGuard)
   @Get()
   async findAll() {
     const data = await this.enrollmentService.findAll();
@@ -27,9 +32,10 @@ export class EnrollmentController {
       data: data,
     };
   }
-
+  @UseGuards(AuthGuard)
   @Get(':id')
-  async findOne(@Param() params: { id: string }) {
+  async findOne(@Param() params: { id: string }, @CurrentUser() user) {
+    if (params.id !== user.id) throw new UnauthorizedException('Acess denied');
     const data = await this.enrollmentService.findOne(params.id);
     return {
       data: data,
@@ -37,16 +43,27 @@ export class EnrollmentController {
   }
 
   @Patch(':id')
-async update(@Body() data, @Param() params:{id:string}){
-const enrollment = await this.enrollmentService.updateEnrollment(params.id, data)
-return {
-  message:"Updated successfuly !",
-  data:enrollment
-}
-}
+  async update(
+    @Body() data,
+    @Param() params: { id: string },
+    @CurrentUser() user,
+  ) {
+    if (params.id !== user.id) throw new UnauthorizedException('Acess denied');
+
+    const enrollment = await this.enrollmentService.updateEnrollment(
+      params.id,
+      data,
+    );
+    return {
+      message: 'Updated successfuly !',
+      data: enrollment,
+    };
+  }
 
   @Delete(':id')
-  async delete(@Param() params: { id: string }) {
+  async delete(@Param() params: { id: string }, @CurrentUser() user) {
+    if (params.id !== user.id) throw new UnauthorizedException('Acess denied');
+
     const data = await this.enrollmentService.delete(params.id);
     return {
       message: 'Unenrolled Successfuly !',
